My_ai_bot/requirements.txt
discord.py==2.1.0
openai==0.26.0
python-dotenv==0.21.0
requests==2.25.1